Implement skip_updateinfos options
This can be used to skip puting any updateinfo meta data to
the media. This may lead into errors when only the latest
version of a package is wanted, but updateinfo is
refering to an older one.

We most likely also need another option to limit the
inclusion of updateinfo's only after a certain point
in time. This would be needed for products being
based on an existing code stream, but are create at
a later point of time.

Issue #14