cli: allow disabling package inclusion when using base images
This commit adds the new build_option base_skip_packages that can be
used in conjunction with base images.
Controls whether packages should be copied in the `/install` directory
when repacking the image.

While it sounds counterproductive, in some specific cases it can
be useful. For SLES, it can be used to ship the Online installer
alongside the rest of the product-composer artifacts, and as a
bonus, it can also share the build number with them.

Signed-off-by: Eugenio Paolantonio <eugenio.paolantonio@suse.com>