cli: keep creating Source and Debug images when using the baseiso feature
If iso creation for Source and Debug images is not explicitly disabled
by the user, keep creating them even if the baseiso feature is enabled.

Only the main media ("Packages") iso would be based on the specified
base iso. Other medias would be created as usual and will not be
bootable.

Signed-off-by: Eugenio Paolantonio <eugenio.paolantonio@suse.com>