productcomposer: createupdateinfoxml: set rpm buildtime by looking at the issued date
If we're skipping packages using a DateFilter, when processing
updateinfos we might die on looking for rpms that were filtered out.

Set the updateinfo issued date as the buildtime in the fake package
entry used for matching. Although it is not precise, it should help in
properly skipping the referenced packages.

Signed-off-by: Eugenio Paolantonio <eugenio.paolantonio@suse.com>